backend/pyproject.toml
[tool.poetry]
name = "ibora-mobi-backend"
version = "0.1.0"
description = "Backend do projeto Ibora Mobi"
authors = ["JP Santos"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.20"
email-validator = "^2.3.0"
bcrypt = "3.2.2"
aiormq = "^6.9.2"
structlog = "^25.5.0"
prometheus-client = "^0.23.1"
opentelemetry-api = "^1.39.1"
opentelemetry-sdk = "^1.39.1"
opentelemetry-instrumentation-fastapi = "^0.60b1"
opentelemetry-exporter-otlp = "^1.39.1"
geoalchemy2 = "^0.18.1"
googlemaps = "^4.10.0"
polyline = "^2.0.4"
gerencianet = "^3.0.0"
apscheduler = "^3.11.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.26.0"
black = "^24.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
